+++ react.js
@@ -1,0 +1,1 @@
+export {default, default as CopyPageButton} from "./src/CopyPageButton";